requirements-lock.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.12.1
attrs==26.1.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
contourpy==1.3.3
coverage==7.13.5
cryptography==46.0.5
cuda-bindings==12.9.4
cuda-pathfinder==1.4.3
cycler==0.12.1
Cython==3.2.4
datasets==4.8.4
Deprecated==1.3.1
dill==0.4.1
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
-e git+ssh://git@github.com/gnai-creator/drm_transformer.git@17c852153012a44cae0e74092cbd56fc80234873#egg=drm_transformer
email-validator==2.3.0
fastapi==0.135.2
filelock==3.25.2
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.2
hopcroftkarp==1.2.5
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.7.2
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
kiwisolver==1.5.0
langdetect==1.0.9
limits==5.8.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
networkx==3.6.1
numpy==2.4.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
openai==2.29.0
packaging==26.0
pandas==3.0.1
pdfminer.six==20251230
pdfplumber==0.11.9
persim==0.3.8
pillow==12.1.1
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.12.1
pyparsing==3.3.2
pypdfium2==5.6.0
pytest==9.0.2
pytest-cov==7.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
ripser==0.6.14
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.8.3
starlette==1.0.0
stripe==14.4.1
sympy==1.14.0
tavily-python==0.7.23
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.3.0
triton==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
